requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
redis==4.6.0
Werkzeug==2.3.7
email-validator==2.0.0
gunicorn==21.2.0